Update testUpdateConfig(): setUp(), tearDown() and refactoring.
