+++ helpers/config.php
@@ -41,30 +41,8 @@
 
         return $content;
     }
-    // Returns the elements of $minuend that are not elements of $subtrahend.
-    // Ignores NULL elements.
-    function array_diff_recursive( $minuend, $subtrahend = [] ) {
-        $difference = [];
-        foreach ( $minuend as $key => $value ) {
-            if ( $value === NULL ) {
-                continue;
-            }
-            if ( is_array( $value ) && $value !== [] ) {
-                if ( !isset( $subtrahend[ $key ] ) ) {
-                    $subtrahend[ $key ] = [];
-                }
-                $result = array_diff_recursive( $value, $subtrahend[ $key ] );
-                if ( $result !== [] ) {
-                    $difference[ $key ] = $result;
-                }
-            }
-            elseif ( !isset( $subtrahend[ $key ] ) || $value !== $subtrahend[ $key ] ) {
-                $difference[ $key ] = $value;
-            }
-        }
-        return $difference;
-    }
-    // Can be used to delete entries that are set to NULL.
+    // Updates $config and the config-local.php file to include the given $entries.
+    // Can also be used to delete $entries that are set to NULL.
     function updateConfig( $entries, $environment ) {
         global $config;
