+++ models/dependencies.php
@@ -8,4 +8,5 @@
     require_once 'helpers/pluralize.php';
     require_once 'helpers/config.php';
     require_once 'helpers/file.php';
+    require_once 'helpers/array.php';
 ?>
